Restructure stdlib into an Acton project
With updates to actonc on how we expect an Acton project to look like
and how we use it to compile the Acton builtin and standard library
itself, it is now easier if the builtins and standard library is
arranged like an Acton project. There is still a lot of magic makefile
voodoo to bring everything together.

numpy is considered part of the stdlib. Since it has a bunch of .c files
etc, it is sort of split up. In stdlib/src we only keep the main module
files, namely:
- numpy.act
- numpy.h
- numpy.c

The other files are in stdlib/c_src/numpy and we include them by having
weirdly long and relative inclusion paths. It's not very pretty at all
but it works. We should probably think a little further about this.

The main Makefile has been further enhanced, in particular around the
building of the distribution, which now entirely consists of declarative
targets with correct dependency definitions.
