--- numpy/numpy.act
@@ -1,45 +1,0 @@
-# Copyright (C) 2019-2021 Data Ductus AB
-#
-# Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
-#
-# 1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer.
-#
-# 2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution.
-#
-# 3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote products derived from this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-#
-
-protocol Primitive: pass
-
-class ndarray[A]: 
-    shape           : list[int]                      # should be a method? (at least until we have immutable lists)
-    strides         : list[int]                      #      -"-
-    reshape         : (list[int]) -> Self
-    transpose       : (?list[int]) -> Self
-    flatten         : () -> Self                     # the same as reshape([-1])
-    copy            : () -> Self                     # makes a contiguous deep copy
-#    __nd_getslice__ : (list[?(int|slice)]) -> Self
-
-
-partition           : [A (Primitive)] => (ndarray[A],int,?int)   -> ndarray[A]
-clip                : [A (Primitive)] => (ndarray[A],?A,?A)      -> ndarray[A]
-dot                 : [A (Primitive)] => (ndarray[A],ndarray[A]) -> ndarray[A]
-sum                 : [A (Primitive)] => (ndarray[A],?int)       -> ndarray[A]
-sort                : [A (Primitive)] => (ndarray[A],?int)       -> ndarray[A]
-abs                 : [A (Primitive)] => (ndarray[A])            -> ndarray[A]
-scalar              : [A (Primitive)] => (ndarray[A])            -> A           #only for 0-dim array
-    
-
-arange              :                    (?int,int,?int)         -> ndarray[int]       
-linspace            :                    (float, float, int)     -> ndarray[float]
-array               : [A (Primitive)] => (list[A])               -> ndarray[A]            
-
-
-extension ndarray[A (Primitive)] (Integral): pass
-
-extension ndarray[A (Primitive)] (Collection[A]): pass
-
-extension int (Primitive): pass
-extension float (Primitive): pass
