--- numpy/numpy.h
@@ -1,9 +1,0 @@
-#include "../builtin/builtin.h"
-
-#include "../numpy/ndselect.h"
-#include "../numpy/primitive.h"
-#include "../numpy/ndarray.h"
-#include "../numpy/protocol_impls.h"
-#include "../numpy/init.h"
-#include "../numpy/quickselect.h"
-
