+++ stdlib/c_src/numpy/numpytests/indextest.c
