+++ stdlib/c_src/numpy/numpytests/opertest.c
