+++ stdlib/c_src/numpy/numpytests/test1.py
