+++ stdlib/c_src/numpy/numpytests/test1_mutate.py
