+++ stdlib/c_src/numpy/primitive.c
