+++ stdlib/c_src/numpy/protocol_impls.c
