+++ stdlib/c_src/numpy/protocol_impls.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "modules/math.h"
+#include "types/math.h"
 
 
 struct numpy$$Integral$ndarray$int;
