+++ stdlib/c_src/numpy/tests/opertest.c
