+++ stdlib/c_src/numpy/tests/strtest.c
