+++ stdlib/src/math.h
