+++ stdlib/src/numpy.act
