+++ stdlib/src/numpy.h
@@ -1,0 +1,8 @@
+#include "builtin/builtin.h"
+
+#include "../c_src/numpy/ndselect.h"
+#include "../c_src/numpy/primitive.h"
+#include "../c_src/numpy/ndarray.h"
+#include "../c_src/numpy/protocol_impls.h"
+#include "../c_src/numpy/init.h"
+#include "../c_src/numpy/quickselect.h"
