+++ stdlib/src/random.act
