+++ stdlib/src/random.h
