+++ stdlib/src/time.act
